Added Particle System, Random other ship spawns,...
Sailing animation swapped for particle effect, in-progress, cause it
doesn't work when moving in the y-axis. Also added more "regions" so
that you can minimize/hide code you're not working on. Should make
reading the code a little easier. Battle Mode, still a work in progress.
Need to work on enemy ship AI in battle mode.