More Save File Refinements
New Changes:
- The player will start a new game in the first open save file available
(if the user for some reason decides to continue a game from the 2nd
file before the 1st file, etc).
- If there are 3 save files already occupied, the user can select which
file to overwrite.

Issues:
- The back button on the continue screen where there are 3 save files
occupied is NOT working correctly, the main menu will load but a
cornflower blue screen is overlaid on top. The main menu buttons are
still functional (even though the user cannot see them)